Updated submit.org
Made it a bit more sane.

Cam up with a concept of a 'task' which is the directory in which the
submission script is run. Then renamed variables to reflect this.

Also simplified the path variables and did away with it being
dependent on your $HOME or whatever. Now the user does that in the
noweb.org file and just gives the preceeding path.

Also moved out the whole environmental variable section to another
noweb insertion point, this makes the script much shorter and comprehendable.